backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.2.0
drf-spectacular==0.27.1
django-filter==23.5
python-dateutil==2.8.2